Added in modular drive controls, added in Utilities class
It's not very well-explained, but by creating a instance of a ControlledMotorGroup command, with an array of ControlledMotorDigital and ControlledMotorAnalog objects one can easily construct a basic drive command.

The parameters of each control element can be modified, such as the output scale, the input controller, and whether deadzones are enabled.

Utilities.java is a class full of static methods used often in robot code. It currently is quite simple, but as the robot code becomes more complex Utilities will as well, allowing future code to be written much faster.